requirements.txt
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=6.0
pandas>=1.3.0
matplotlib>=3.5.0